backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.29
alembic==1.13.1
psycopg2-binary==2.9.9
pydantic[email]==2.6.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.22
jinja2==3.1.3
openpyxl==3.1.2
qrcode[pil]==7.4.2
python-dotenv==1.0.1
pytest==8.1.1
httpx==0.27.0
